functions/requirements.txt
firebase-functions==0.4.2
firebase-admin==6.6.0
requests==2.31.0
functions-framework==3.*

google-cloud-storage==2.*
google-cloud-pubsub==2.*
httpx>=0.27.0
fastapi>=0.100.0 # Changed from 0.68.* for Pydantic v2 compatibility
uvicorn==0.15.*

# Flask to serve HTTP requests.
Flask==2.2.5

# OpenAI Python client library to call the ChatGPT API.
openai==1.61.1
pydantic
Pillow
uuid
anthropic>=0.7.0
moondream

google-genai >= 0.1.0